Move utils out of libs
