+++ archetypes/ArbitrumBridge/Modal.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { TWModal } from '@components/General/TWModal';
-import { toApproxCurrency } from '@libs/utils/converters';
+import { toApproxCurrency } from '~/utils/converters';
 import { useReducer } from 'react';
 import { bridgeReducer, DefaultBridgeState, BridgeStepEnum } from './state';
