+++ archetypes/Exchange/Inputs/index.tsx
@@ -4,7 +4,7 @@ import { CommitActionEnum, BalanceTypeEnum, SideEnum } from '@tracer-protocol/po
 import TWButtonGroup from '@components/General/TWButtonGroup';
 import { PoolInfo } from '@context/PoolContext/poolDispatch';
 import { SwapState, useBigNumber, SwapAction } from '@context/SwapContext';
-import { toApproxCurrency } from '@libs/utils/converters';
+import { toApproxCurrency } from '~/utils/converters';
 import usePoolTokens from '~/hooks/usePoolTokens';
 import usePoolsNextBalances from '~/hooks/usePoolsNextBalances';
 import TokenSelect from '../TokenSelect';
