+++ archetypes/Exchange/Summary/ApproxCommitGasFee.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import BigNumber from 'bignumber.js';
-import { toApproxCurrency } from '@libs/utils/converters';
+import { toApproxCurrency } from '~/utils/converters';
 import { Transparent } from './styles';
 
 export const ApproxCommitGasFee: React.FC<{
