+++ archetypes/Pools/PoolDetailsModal/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { TWModal } from '@components/General/TWModal';
 import { Table, TableRow, TableRowCell } from '@components/General/TWTable';
 import styled from 'styled-components';
-import { getPriceFeedUrl } from '@libs/utils/converters';
+import { getPriceFeedUrl } from '~/utils/converters';
 import { Theme } from '@store/ThemeSlice/themes';
 
 import FollowLink from '/public/img/general/follow-link.svg';
