+++ archetypes/Pools/PoolsTable/index.tsx
@@ -3,7 +3,7 @@ import { CommitActionEnum, SideEnum } from '@tracer-protocol/pools-js';
 import Button from '@components/General/Button';
 import { Table, TableHeader, TableRow, TableHeaderCell, TableRowCell } from '@components/General/TWTable';
 import { ARBITRUM } from '~/constants/networks';
-import { calcPercentageDifference, getPriceFeedUrl, toApproxCurrency } from '@libs/utils/converters';
+import { calcPercentageDifference, getPriceFeedUrl, toApproxCurrency } from '~/utils/converters';
 import { BrowseTableRowData, DeltaEnum } from '../state';
 import { TWModal } from '@components/General/TWModal';
 import TimeLeft from '@components/TimeLeft';
@@ -11,7 +11,7 @@ import Actions from '@components/TokenActions';
 import { Logo, LogoTicker, tokenSymbolToLogoTicker } from '@components/General';
 import { useWeb3 } from '@context/Web3Context/Web3Context';
 import { ethers } from 'ethers';
-import { ArbiscanEnum } from '@libs/utils/rpcMethods';
+import { ArbiscanEnum } from '~/utils/rpcMethods';
 import TooltipSelector, { TooltipKeys } from '@components/Tooltips/TooltipSelector';
 import useIntervalCheck from '~/hooks/useIntervalCheck';
 import { LinkOutlined } from '@ant-design/icons';
@@ -20,7 +20,7 @@ import styled from 'styled-components';
 import { Theme } from '@store/ThemeSlice/themes';
 
 import Close from '/public/img/general/close.svg';
-import { classNames } from '@libs/utils/functions';
+import { classNames } from '~/utils/helpers';
 import { constructBalancerLink } from '@archetypes/BalancerBuySell';
 import { StyledTooltip } from '@components/Tooltips';
 import { default as UpOrDown } from '@components/UpOrDown';
