+++ archetypes/Pools/index.tsx
@@ -16,7 +16,7 @@ import { useWeb3 } from '@context/Web3Context/Web3Context';
 import useBrowsePools from '~/hooks/useBrowsePools';
 import { noDispatch, useSwapContext } from '@context/SwapContext';
 import MintBurnModal from './MintBurnModal';
-import { marketFilter } from '@libs/utils/functions';
+import { marketFilter } from '~/utils/filters';
 import Loading from '@components/General/Loading';
 
 export const Browse: React.FC = () => {
