+++ components/General/Input/Numeric/index.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback, KeyboardEventHandler } from 'react';
-import { classNames } from '@libs/utils/functions';
+import { classNames } from '~/utils/helpers';
 
 // const inputRegex = RegExp(`^\\d*(?:\\\\[.])?\\d*$`); // match escaped "." characters via in a non-capturing group
