+++ components/General/Notification/Commit.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useRouter } from 'next/router';
 import { ethers } from 'ethers';
 import TimeLeft from '@components/TimeLeft';
-import { watchAsset } from '@libs/utils/rpcMethods';
+import { watchAsset } from '~/utils/rpcMethods';
 import { CommitProps } from '@store/TransactionSlice/types';
 import { Notification } from '.';
 import { Logo, tokenSymbolToLogoTicker } from '..';
