+++ components/General/TWTable/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { Theme } from '@store/ThemeSlice/themes';
-import { classNames } from '@libs/utils/functions';
+import { classNames } from '~/utils/helpers';
 
 export const Table: React.FC<{ showDivider?: boolean; className?: string }> = ({
     showDivider = false,
