+++ components/General/UnsupportedNetworkPopup/index.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect } from 'react';
 import { toast } from 'react-toastify';
 import { useWeb3 } from '@context/Web3Context/Web3Context';
-import { isSupportedNetwork, isSupportedBridgeNetwork } from '@libs/utils/supportedNetworks';
+import { isSupportedNetwork, isSupportedBridgeNetwork } from '~/utils/supportedNetworks';
 import { ARBITRUM } from '~/constants/networks';
-import { switchNetworks } from '@libs/utils/rpcMethods';
+import { switchNetworks } from '~/utils/rpcMethods';
 import { useRouter } from 'next/router';
 import { Notification } from '../Notification';
