+++ components/General/index.tsx
@@ -1,4 +1,4 @@
-import { classNames } from '@libs/utils/functions';
+import { classNames } from '~/utils/helpers';
 import { Children } from '~/types/general';
 import React from 'react';
 import styled from 'styled-components';
