+++ components/Nav/Navbar/AccountBalance.tsx
@@ -2,8 +2,8 @@ import React, { useState, useEffect } from 'react';
 import { BigNumber } from 'bignumber.js';
 import { useStore } from '@store/main';
 import { usePools } from '@context/PoolContext';
-import { toApproxCurrency } from '@libs/utils/converters';
-import { classNames } from '@libs/utils/functions';
+import { toApproxCurrency } from '~/utils/converters';
+import { classNames } from '~/utils/helpers';
 import TracerLoading from '@public/img/logos/tracer/tracer-loading-white.svg';
 import { selectPendingCount } from '@store/TransactionSlice';
