+++ components/Nav/Navbar/NetworkDropdown.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Logo, LogoTicker } from '@components/General';
 import { useWeb3 } from '@context/Web3Context/Web3Context';
-import { switchNetworks } from '@libs/utils/rpcMethods';
+import { switchNetworks } from '~/utils/rpcMethods';
 import { networkConfig } from '@context/Web3Context/Web3Context.Config';
 import { ARBITRUM, ARBITRUM_RINKEBY } from '~/constants/networks';
 import TWPopup from '@components/General/TWPopup';
