+++ utils/filters.ts
@@ -1,13 +1,5 @@
 import { MarketFilterEnum } from '~/types/filters';
 
-export function classNames(...classes: string[]): string {
-    return classes.filter(Boolean).join(' ');
-}
-
-export function escapeRegExp(string: string): string {
-    return string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&'); // $& means the whole matched string
-}
-
 export const marketFilter: (poolName: string, marketFilterState: MarketFilterEnum) => boolean = (
     poolName,
     marketFilterState,
