+++ utils/helpers.ts
@@ -1,0 +1,3 @@
+export function classNames(...classes: string[]): string {
+    return classes.filter(Boolean).join(' ');
+}
